added tests

+++ src/main/java/com/Gydytojaidiagnozes/web/service/DoctorService.java
@@ -17,22 +17,6 @@ public class DoctorService {
     @Autowired
     DiagnoseRepository diagnoseRepository;
 
-    /*private final DoctorRepository doctorRepository;
-    private final DiagnoseRepository diagnoseRepository;*/
-
-    private static List<Doctor> doctors = new ArrayList<>();
-    private static int doctorsCount = 3;
-
-    static {
-        doctors.add(new Doctor(1, "Petras", "+37062742568"));
-        doctors.add(new Doctor(2, "Jonas", "+37062742788"));
-        doctors.add(new Doctor(3, "Gintaras", "+37062522568"));
-    }
-
-    /*public DoctorService(DoctorRepository doctorRepository, DiagnoseRepository diagnoseRepository) {
-        this.doctorRepository = doctorRepository;
-        this.diagnoseRepository = diagnoseRepository;
-    }*/
 
     public Doctor findDoctorById(int id) {
         return doctorRepository.findById(id);
@@ -57,7 +41,6 @@ public class DoctorService {
     public void updateDoctor(Doctor doctor) {
         Doctor updatable = doctorRepository.findById(doctor.getId());
         if(updatable != null) {
-            System.out.println(updatable.getName());
             updatable.setName(doctor.getName());
             updatable.setTelephoneNumber(doctor.getTelephoneNumber());
             doctorRepository.save(updatable);
@@ -68,17 +51,18 @@ public class DoctorService {
     }
 
     public void deleteDoctorById(int id) {
-        doctorRepository.deleteById(id);
-        diagnoseRepository.deleteAllByDoctorId(id);
+        Doctor doctor = doctorRepository.findById(id);
+        if(doctor != null) {
+            doctorRepository.deleteById(id);
+            if (diagnoseRepository.findByDoctorId(doctor) != null) diagnoseRepository.deleteAllByDoctorId(doctor);
+        }
     }
 
     public void deleteDoctor(Doctor doctor) {
         doctorRepository.delete(doctor);
-        diagnoseRepository.deleteAllByDoctorId(doctor.getId());
+        diagnoseRepository.deleteAllByDoctorId(doctor);
     }
-    public List<Doctor> retrieveDoctors() {
-        return doctors;
-    }
+
 }
